refactor: split large utilities.ts into focused modules
- Create component-utilities.ts for React component types (216 lines)
- Create form-utilities.ts for form and data types (78 lines)
- Create api-utilities.ts for API and chart types (52 lines)
- Create ui-utilities.ts for styling and UI types (327 lines)
- Update utilities.ts to re-export from focused modules (132 lines)

This addresses the Copilot PR review feedback about the 668-line
utilities.ts file being too large and difficult to maintain.
The new structure improves discoverability and maintainability
while preserving backward compatibility through re-exports.
